Allows to specify a domain when deleting a cookie.
When a cookie has been set with a given domain (i.e. domain="example.com"), subsequent attempts to delete it will be ignored by the browser if the same domain is not provided.

So a 'domain' parameter is added to the 'delete_cookie' method in order to be able to delete those cookies.